Move macro expand session related codes to new module
Add "Language.Finkel.Make.Session", move the codes used for managing the HscEnv
session for macro expansion from "Language.Finkel.Expand". Update the module
importing the moved codes.

Some minor comment updates.
