+++ finkel-kernel/finkel-kernel.cabal
@@ -74,6 +74,7 @@ library
                        Language.Finkel.Data.SourceText
                        Language.Finkel.Make.Cache
                        Language.Finkel.Make.Recompile
+                       Language.Finkel.Make.Session
                        Language.Finkel.Make.Summary
                        Language.Finkel.Make.TargetSource
                        Language.Finkel.Make.Trace
